backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
Pillow
numpy
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
httpx
python-dotenv
